Updated scripts

+++ deploy.sh
@@ -1,22 +1,37 @@
 #!/bin/bash
 
-APP_NAME=youtube2podcast
-APP_USER=web
+set -x
 
-TARGET_DIR=/usr/web/youtube2podcast
+
+TARGET_DIR="$1"
+APP_USER="$2"
+
+if [ ! "$TARGET_DIR" ] || [ ! "$APP_USER" ]; then
+  echo "Usage: $0 target_dir user_name"
+  exit 1
+fi
+
+
+cd $BASEDIR
+BASEDIR=$PWD
+
+APP_NAME=$(basename $BASEDIR)
+BASEDIR=$(dirname $0)
+SOURCE_JAR=$BASEDIR/build/libs/$APP_NAME.jar
+if [ ! -f "$SOURCE_JAR" ]; then
+  echo "File '$SOURCE_JAR' not found! Needs to build before deploying"
+  exit 1;
+fi
+
+if [ ! "$(echo $TARGET_DIR | grep $APP_NAME)" ]; then
+  TARGET_DIR=$TARGET_DIR/$APP_NAME
+fi
+
 TARGET_JAR=$TARGET_DIR/build/libs/$APP_NAME.jar
 
-BASEDIR=$(dirname $0)
 cp -rvf $BASEDIR/ $TARGET_DIR
-sudo chown -R $APP_USER:$APP_USER $TARGET_DIR
+chown -R $APP_USER:$APP_USER $TARGET_DIR
 
-cd $TARGET_DIR
-
-rm -rf $TARGET_JAR
-sudo -u $APP_USER ./gradlew build
-if [ "$?" != "0" ]; then exit $?; fi
-
-cp -v ./build/libs/$APP_NAME.jar $TARGET_JAR
 rm -f /etc/init.d/$APP_NAME
 ln -s $TARGET_JAR /etc/init.d/$APP_NAME
